Replaces codegansta/cli with cobra/viper:
Environment variable support is broken in codegansta/cli and given
the number of outstanding pull requests on the repo I figure that
project should be consider unmaintained.  This commit replaces it
with cobra and viper, fixing environment variable support.

In addition, I add a simple .env file for ease of development.  Use
something like dotenv of godotenv to use it.