Refine logging:
This commit removes the context argument from the logging commands.
Instead, a new method `Ctx()` is used to retrieve the logger for a
context when available, but otherwise a package can use the `log`
package methods (which no longer need a context) to log when no context
is readily available.

Additionally, this commit fully wraps the logrus library.  There is no
more leakage as logrus.Fields has been wrapped with the log.F type, and
logrus.Logger has been wrapped with log.Logger.