Refactor importation system, introducing the Cursor and Ingestion types
This commit greatly refactors the entire ingestion system.  With this
commit, the major portions of the system become the Session, the
Cursor, and the Ingestion.  The Cursor manages the iteration through
data loaded from stellar-core database.  The Ingestion manages writing
batches of data to the horizon database, and the Session manages the
interaction between the two.