Changes ingestion system to auto-increment for history_accounts:
This commit changes the ingestion system such that records in the
`history_accounts` table are no longer identified by the operation in
which the account was created, but rather they are now simply numbered
automatically using a postgres sequence.  In addition a bug was
discovered while implementing these changes in the “participants”
package and has been fixed in this commit.